# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для системы управления событиями EventFlow",
        "title": "EventFlow API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@eventflow.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/tickets": {
            "get": {
                "description": "Возвращает список всех тикетов с пагинацией",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Получить список тикетов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Пагинация [start, end]",
                        "name": "range",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Сортировка [field, order]",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Ticket"
                            }
                        },
                        "headers": {
                            "Content-Range": {
                                "type": "string",
                                "description": "Диапазон записей"
                            },
                            "X-Total-Count": {
                                "type": "string",
                                "description": "Общее количество записей"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Создает новый тикет с уникальным QR-кодом",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Создать тикет",
                "parameters": [
                    {
                        "description": "Данные тикета",
                        "name": "ticket",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateTicketRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Ticket"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tickets/qr/{qrcode}": {
            "get": {
                "description": "Получение тикета по QR-коду для отслеживания посещаемости",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Получить тикет по QR-коду",
                "parameters": [
                    {
                        "type": "string",
                        "description": "QR-код тикета",
                        "name": "qrcode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Ticket"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tickets/qr/{qrcode}/use": {
            "post": {
                "description": "Отметить тикет как использованный для отслеживания посещаемости",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Отметить тикет как использованный",
                "parameters": [
                    {
                        "type": "string",
                        "description": "QR-код тикета",
                        "name": "qrcode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Ticket"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.CreateTicketRequest": {
            "type": "object",
            "required": [
                "event_id",
                "participant_id",
                "status",
                "ticket_type"
            ],
            "properties": {
                "event_id": {
                    "type": "integer"
                },
                "participant_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "ticket_type": {
                    "type": "string"
                }
            }
        },
...
        "models.Ticket": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "event_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "participant_id": {
                    "type": "integer"
                },
                "qr_code": {
                    "type": "string"
                },
                "status": {
                    "description": "\"active\" или \"canceled\"",
                    "type": "string"
                },
                "ticket_type": {
                    "description": "\"free\" или \"paid\"",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}